libredirect: fix type of listxattr, llistxattr func pointers
Copy paste mistake.

+++ pkgs/by-name/li/libredirect/libredirect.c
@@ -390,7 +390,7 @@ WRAPPER_DEF(opendir)
 #if !defined(__APPLE__)
 WRAPPER(ssize_t, listxattr)(const char * path, char * list, size_t size)
 {
-    int (*listxattr_real) (const char *, char *, size_t) = LOOKUP_REAL(listxattr);
+    ssize_t (*listxattr_real) (const char *, char *, size_t) = LOOKUP_REAL(listxattr);
     char buf[PATH_MAX];
     return listxattr_real(rewrite(path, buf), list, size);
 }
@@ -400,7 +400,7 @@ WRAPPER_DEF(listxattr);
 #if !defined(__APPLE__)
 WRAPPER(ssize_t, llistxattr)(const char * path, char * list, size_t size)
 {
-    int (*llistxattr_real) (const char *, char *, size_t) = LOOKUP_REAL(llistxattr);
+    ssize_t (*llistxattr_real) (const char *, char *, size_t) = LOOKUP_REAL(llistxattr);
     char buf[PATH_MAX];
     return llistxattr_real(rewrite(path, buf), list, size);
 }
